Outsource getter helper
